add a vector based ctor for perf stats
